refactor: 상품 및 쿠폰 데이터를 별도 상수 파일로 분리
