+++ src/basic/App.tsx
@@ -1,11 +1,8 @@
 import { useState, useCallback, useEffect } from 'react';
 import { CartItem, Coupon, Product } from '../types';
 import { formatPrice } from './utils/formatters';
-
-interface ProductWithUI extends Product {
-  description?: string;
-  isRecommended?: boolean;
-}
+import { Products } from './constants/products';
+import { Coupons } from './constants/coupons';
 
 interface Notification {
   id: string;
@@ -13,67 +10,17 @@ interface Notification {
   type: 'error' | 'success' | 'warning';
 }
 
-// 초기 데이터
-const initialProducts: ProductWithUI[] = [
-  {
-    id: 'p1',
-    name: '상품1',
-    price: 10000,
-    stock: 20,
-    discounts: [
-      { quantity: 10, rate: 0.1 },
-      { quantity: 20, rate: 0.2 },
-    ],
-    description: '최고급 품질의 프리미엄 상품입니다.',
-  },
-  {
-    id: 'p2',
-    name: '상품2',
-    price: 20000,
-    stock: 20,
-    discounts: [{ quantity: 10, rate: 0.15 }],
-    description: '다양한 기능을 갖춘 실용적인 상품입니다.',
-    isRecommended: true,
-  },
-  {
-    id: 'p3',
-    name: '상품3',
-    price: 30000,
-    stock: 20,
-    discounts: [
-      { quantity: 10, rate: 0.2 },
-      { quantity: 30, rate: 0.25 },
-    ],
-    description: '대용량과 고성능을 자랑하는 상품입니다.',
-  },
-];
-
-const initialCoupons: Coupon[] = [
-  {
-    name: '5000원 할인',
-    code: 'AMOUNT5000',
-    discountType: 'amount',
-    discountValue: 5000,
-  },
-  {
-    name: '10% 할인',
-    code: 'PERCENT10',
-    discountType: 'percentage',
-    discountValue: 10,
-  },
-];
-
 const App = () => {
-  const [products, setProducts] = useState<ProductWithUI[]>(() => {
+  const [products, setProducts] = useState<typeof Products>(() => {
     const saved = localStorage.getItem('products');
     if (saved) {
       try {
         return JSON.parse(saved);
       } catch {
-        return initialProducts;
+        return Products;
       }
     }
-    return initialProducts;
+    return Products;
   });
 
   const [cart, setCart] = useState<CartItem[]>(() => {
@@ -88,16 +35,16 @@ const App = () => {
     return [];
   });
 
-  const [coupons, setCoupons] = useState<Coupon[]>(() => {
+  const [coupons, setCoupons] = useState<typeof Coupons>(() => {
     const saved = localStorage.getItem('coupons');
     if (saved) {
       try {
         return JSON.parse(saved);
       } catch {
-        return initialCoupons;
+        return Coupons;
       }
     }
-    return initialCoupons;
+    return Coupons;
   });
 
   const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
@@ -239,7 +186,7 @@ const App = () => {
   }, [searchTerm]);
 
   const addToCart = useCallback(
-    (product: ProductWithUI) => {
+    (product: (typeof Products)[number]) => {
       const remainingStock = getRemainingStock(product);
       if (remainingStock <= 0) {
         addNotification('재고가 부족합니다!', 'error');
@@ -322,8 +269,8 @@ const App = () => {
   }, [addNotification]);
 
   const addProduct = useCallback(
-    (newProduct: Omit<ProductWithUI, 'id'>) => {
-      const product: ProductWithUI = {
+    (newProduct: Omit<(typeof Products)[number], 'id'>) => {
+      const product: (typeof Products)[number] = {
         ...newProduct,
         id: `p${Date.now()}`,
       };
@@ -334,7 +281,7 @@ const App = () => {
   );
 
   const updateProduct = useCallback(
-    (productId: string, updates: Partial<ProductWithUI>) => {
+    (productId: string, updates: Partial<(typeof Products)[number]>) => {
       setProducts((prev) =>
         prev.map((product) => (product.id === productId ? { ...product, ...updates } : product)),
       );
@@ -403,7 +350,7 @@ const App = () => {
     setShowCouponForm(false);
   };
 
-  const startEditProduct = (product: ProductWithUI) => {
+  const startEditProduct = (product: (typeof Products)[number]) => {
     setEditingProduct(product.id);
     setProductForm({
       name: product.name,
